notify_debouncer_mini

Function new_debouncer

pub fn new_debouncer<F: DebounceEventHandler>(
    timeout: Duration,
    event_handler: F,
) -> Result<Debouncer<RecommendedWatcher>, Error>

Short function to create a new debounced watcher with the recommended debouncer.

Timeout is the amount of time after which a debounced event is emitted or a continuous event is send, if there still are events incoming for the specific path.